backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.6.15
distro==1.9.0
Django==5.2.3
djangorestframework==3.16.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.86.0
psycopg==3.2.9
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
python-dotenv==1.0.1
requests==2.32.3

celery==5.3.6
redis==5.0.4
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
pytest==8.2.2
pytest-django==4.8.0
